chore: Update dependencies in Cargo.toml and package.json
Updated the dependencies in the Cargo.toml and package.json files to their latest versions. This includes upgrading tauri-build to version 1.5.4, serde_json to version 1.0.128, serde to version 1.0.210, tauri to version 1.7.2, winreg to version 0.52.0, wmi to version 0.14.0, and other dependencies in the package.json file. These updates ensure that the project is using the most recent versions of the dependencies, improving stability and security.